split the host class into a Scraper superclass and the Host subclass
